+++ src/pages/FlowRun/Restart-Dialog.vue
@@ -6,6 +6,14 @@ export default {
     flowRun: {
       required: true,
       type: Object
+    },
+    eligibleStates: {
+      required: true,
+      type: Array
+    },
+    failedTaskRuns: {
+      required: true,
+      type: Array
     }
   },
   data() {
@@ -19,11 +27,19 @@ export default {
     restartMessage() {
       return `${this.user.username} restarted this flow run`
     },
-    isFailedRun() {
-      return this.flowRun.state == 'Failed' || this.hasFailedTaskRuns
+    isEligibleToRestart() {
+      return (
+        this.hasFailedTaskRuns ||
+        this.eligibleStates.includes(this.flowRun.state)
+      )
     },
     hasFailedTaskRuns() {
       return this.failedTaskRuns?.length > 0
+    },
+    reRunStates() {
+      const reRunStates = this.eligibleStates.slice()
+      reRunStates.splice(this.eligibleStates.length - 1, 0, '&')
+      return reRunStates.join(' ')
     }
   },
   methods: {
@@ -74,7 +90,7 @@ export default {
 
         if (
           result?.data?.set_task_run_states ||
-          this.flowRun.state == 'Failed'
+          this.eligibleStates.includes(this.flowRun.state)
         ) {
           const { data } = await this.$apollo.mutate({
             mutation: require('@/graphql/TaskRun/set-flow-run-states.gql'),
@@ -124,15 +140,6 @@ export default {
     }
   },
   apollo: {
-    failedTaskRuns: {
-      query: require('@/graphql/FlowRun/failed-task-runs.gql'),
-      variables() {
-        return {
-          flowRunId: this.flowRun.id
-        }
-      },
-      update: data => data?.task_run
-    },
     utilityDownstreamTasks: {
       query: require('@/graphql/TaskRun/utility_downstream_tasks.gql'),
       variables() {
@@ -156,37 +163,25 @@ export default {
 
 <template>
   <v-card tile>
-    <v-card-title>
-      Restart from failed?
-    </v-card-title>
+    <v-card-title> Restart? </v-card-title>
 
     <v-card-text>
-      Click on confirm to restart
-      <span class="font-weight-bold">{{ flowRun.name }}</span>
+      Click on confirm to restart this flow run. Restarting will restart all
+      task runs in {{ reRunStates }} states and run any task runs in a Pending
+      state. Task runs that rely on upstream results may require additional
+      configuration to restart correctly. Read more about that
+      <a
+        href="https://docs.prefect.io/core/concepts/results.html#results"
+        target="_blank"
+        >in the Results docs</a
+      >.
     </v-card-text>
     <v-card-actions>
       <v-spacer></v-spacer>
 
-      <v-tooltip bottom>
-        <template v-slot:activator="{ on }">
-          <div v-on="on">
-            <v-btn
-              :disabled="!isFailedRun"
-              color="primary"
-              @click="restart"
-              v-on="on"
-            >
-              Confirm
-            </v-btn>
-          </div>
-        </template>
-        <span v-if="role === 'READ_ONLY_USER'">
-          Read-only users cannot restart flow runs
-        </span>
-        <span v-else-if="!isFailedRun">
-          You can only restart a failed flow run.
-        </span>
-      </v-tooltip>
+      <v-btn :disabled="!isEligibleToRestart" color="primary" @click="restart">
+        Confirm
+      </v-btn>
 
       <v-btn text @click="cancel">
         Cancel